added alternative method to parse fileobj
depending on the type of fileobj, lxml.html.parse might return an
error because it is expecting a file object instead that a string
(see http://stackoverflow.com/a/8917972 for more info).

Insted, we try to parse the string with lxml.etree.fromstring first,
fallbacking to the original method if it throws a ValueError exception.